requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
python-jose[cryptography]